로그인 설정

+++ src/main/java/com/prj1/controller/BoardController.java
@@ -3,11 +3,14 @@ package com.prj1.controller;
 import com.prj1.domain.Board;
 import com.prj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 
 @Controller
 @RequiredArgsConstructor
@@ -21,19 +24,17 @@ public class BoardController {
     }
 
     @PostMapping("/add")
-    public String addPost(Board board, RedirectAttributes rttr) {
+    public String addPost(Board board, Authentication authentication, RedirectAttributes rttr) {
         System.out.println("board = " + board);
-        service.add(board);
+        service.add(board, authentication);
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
     }
 
     @GetMapping("/board")
     public String view(Integer id, Model model) {
-        // 게시물 조회(select)
+        // 글 작성자 게시물 조회(select)
         Board board = service.get(id);
-
-        System.out.println("board = " + board);
 
         // 모델에 넣고
         model.addAttribute("board", board);
@@ -42,18 +43,22 @@ public class BoardController {
     }
 
     @GetMapping("/")
-    public String home(Model model) {
+    public String home(@RequestParam(value = "page", defaultValue = "1") Integer page, Model model) {
         // 게시물 목록 조회(select)
         service.list();
-        // 모델에 넣고
-        model.addAttribute("boardList", service.list());
+        // 모델에 넣고 . 페이징 하기 전
+//        model.addAttribute("boardList", service.list(page));
+        model.addAllAttributes(service.list(page));
         // jsp로 포워드
         return "board/home";
     }
 
     @PostMapping("/delete")
-    public String delete(Integer id) {
-        service.remove(id);
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
+
         return "redirect:/";
     }
 
@@ -67,8 +72,11 @@ public class BoardController {
     }
 
     @PostMapping("/modify")
-    public String modifyPost(Board board, RedirectAttributes rttr) {
-        service.modify(board);
+    public String modifyPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.modify(board);
+        }
+
         System.out.println("board = " + board);
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
